backend/requirements.txt
websockets>=11.0
webrtcvad-wheels
elevenlabs
openai
python-dotenv
aiohttp
numpy
scipy
noisereduce